Added basic react-for-role feature
Implemented a basic map in the format {guild -> {message -> role, ...}, ...}, where a map of guilds is associated with maps of messages to roles. For example, guild ServerA will be associated to a map where each entry in that map is a message that, when reacted to, results in the reacting user being granted the role associated to that message.
An example is currently hardcoded for personal use, but a more verbose and customizable system will be implemented in the future