Removed use of CommandPrivilege
It's not reflected in the API yet, but Discord Permissions V2 has given server admins the ability to control the command privileges for their servers. Bots are no longer able to use the same endpoint.
Applications that wish to continue exercising this power have to use the OAuth2 method and be given permission by server admins